qt: Fix build against Qt 5.9
* lang/qt/tests/t-keylist.cpp (KeyListTest::testListAllKeysSync):
Ensure same type for both arguments of QCOMPARE.
* lang/qt/tests/t-various.cpp (TestVarious::testSetExpire):
Ensure same type for both arguments of QCOMPARE.
--

Qt 5.9 does not yet have the generic qCompare helper which supports
arguments of any type that can be compared with ==.

GnuPG-bug-id: 5592

+++ lang/qt/tests/t-various.cpp
@@ -230,7 +230,7 @@ private Q_SLOTS:
         std::vector<Subkey> primaryKey;
         primaryKey.push_back(key.subkey(0));
         const auto err = ctx->setExpire(key, 3000, primaryKey);
-        QCOMPARE(err.code(), GPG_ERR_NOT_FOUND);
+        QCOMPARE(err.code(), static_cast<int>(GPG_ERR_NOT_FOUND));
         delete ctx;
     }
